+++ services/madoc-ts/src/frontend/admin/pages/content/canvases/canvas-search-index.tsx
@@ -81,7 +81,7 @@ export const CanvasSearchIndex = createUniversalComponent<CanvasSearchIndexType>
     getData: async (key, { id }, api) => {
       try {
         return {
-          canvas: await api.searchGetIIIF(`urn:madoc:canvas:${id}`),
+          canvas: await api.search.searchGetIIIF(`urn:madoc:canvas:${id}`),
           models: { results: [] },
           // models: await api.searchListModels({ iiif__madoc_id: `urn:madoc:canvas:${id}` }),
         };
