+++ services/madoc-ts/src/routes/iiif/canvases/delete-canvas.ts
@@ -37,7 +37,7 @@ export async function deleteCanvas(canvasId: number, siteId: number, connection:
   });
 
   if (fullyDelete && deletionSummary.search.indexed && deletionSummary.search.id) {
-    await siteApi.searchDeleteIIIF(deletionSummary.search.id);
+    await siteApi.search.searchDeleteIIIF(deletionSummary.search.id);
   }
 
   if (deletionSummary.tasks > 0 || deletionSummary.parentTasks > 0) {
