+++ services/madoc-ts/src/routes/iiif/manifests/delete-manifest-summary.ts
@@ -37,7 +37,7 @@ export async function buildManifestDeletionSummary(
   );
 
   // Search
-  const iiifSearchItem = await siteApi.searchGetIIIF(`urn:madoc:manifest:${manifestId}`).catch(() => null);
+  const iiifSearchItem = await siteApi.search.searchGetIIIF(`urn:madoc:manifest:${manifestId}`).catch(() => null);
 
   // Tasks
   const tasks = await siteApi.getTasks(0, {
