+++ services/madoc-ts/src/routes/search/index-canvas.ts
@@ -73,7 +73,7 @@ export const indexCanvas: RouteMiddleware<{ id: string }> = async context => {
     ]
   );
 
-  await userApi.enrichmentIngestResource(searchPayload);
+  await userApi.search.enrichmentIngestResource(searchPayload);
 
   // 1. Load all capture models for this canvas on this site.
   const models = await userApi.getAllCaptureModels({
